requirements.txt
# Use: python -m pip install -r requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
bbrest==0.3.9
boto3==1.14.2
botocore==1.17.2
certifi==2020.4.5.2
chardet==3.0.4
dateparser==0.7.6
docutils==0.15.2
humanize==2.4.0
idna==2.9
jmespath==0.10.0
maya==0.6.1
multidict==4.7.6
oauthlib==3.1.0
pendulum==2.1.0
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2019.3
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snaptime==0.2.4
tzlocal==2.1
urllib3==1.25.9
yarl==1.4.2
cachetools==4.0.0
PyJWT==1.7.1